refactor(peagen): relocate evaluators under plugins
