--- pkgs/standards/peagen/peagen/evaluators/__init__.py
@@ -1,24 +1,0 @@
-"""Peagen evaluators."""
-
-from .base import Evaluator
-from .psutil_io import PsutilIOEvaluator
-from .pytest_perf_regression import PytestPerfRegressionEvaluator
-
-from .pytest_monitor import PytestMonitorEvaluator
-from .pytest_memray_evaluator import PytestMemrayEvaluator
-from .pytest_profiling import PytestProfilingEvaluator
-
-from .benchmark import PytestBenchmarkEvaluator
-from .simple_time import SimpleTimeEvaluator
-
-
-__all__ = [
-    "Evaluator",
-    "PytestBenchmarkEvaluator",
-    "SimpleTimeEvaluator",
-    "PytestProfilingEvaluator",
-    "PytestMonitorEvaluator",
-    "PytestMemrayEvaluator",
-    "PytestPerfRegressionEvaluator",
-    "PsutilIOEvaluator",
-]
