+++ pkgs/standards/peagen/peagen/plugins/evaluators/__init__.py
@@ -1,8 +1,27 @@
+"""Built-in evaluator plugins."""
+
+from .base import Evaluator
+from .psutil_io import PsutilIOEvaluator
+from .pytest_perf_regression import PytestPerfRegressionEvaluator
+from .pytest_monitor import PytestMonitorEvaluator
+from .pytest_memray_evaluator import PytestMemrayEvaluator
+from .pytest_profiling import PytestProfilingEvaluator
+from .benchmark import PytestBenchmarkEvaluator
+from .simple_time import SimpleTimeEvaluator
+
 from .performance_evaluator import PerformanceEvaluator
 from .ruff_evaluator import RuffEvaluator
 from .pytest_evaluator import PytestEvaluator
 
 __all__ = [
+    "Evaluator",
+    "PytestBenchmarkEvaluator",
+    "SimpleTimeEvaluator",
+    "PytestProfilingEvaluator",
+    "PytestMonitorEvaluator",
+    "PytestMemrayEvaluator",
+    "PytestPerfRegressionEvaluator",
+    "PsutilIOEvaluator",
     "PerformanceEvaluator",
     "RuffEvaluator",
     "PytestEvaluator",
